refactor: move RecipeList out of Workshop
