+++ src/components/Workshop/ForgeTabPanel.js
@@ -7,13 +7,11 @@ import Divider from '@material-ui/core/Divider'
 import ReactMarkdown from 'react-markdown'
 
 import { recipeType } from '../../enums'
-
 import { recipeCategories } from '../../data/maps'
-
 import UpgradePurchase from '../UpgradePurchase'
+import { RecipeList } from '../RecipeList/RecipeList'
 
 import { TabPanel } from './TabPanel'
-import { RecipeList } from './RecipeList'
 
 import { getUpgradesAvailable } from './getUpgradesAvailable'
 
@@ -21,7 +19,6 @@ export function ForgeTabPanel({
   currentTab,
   index,
   learnedForgeRecipes,
-  setCurrentTab,
   toolLevels,
 }) {
   const upgradesAvailable = getUpgradesAvailable({
